requirements.txt
anyio==3.3.2
boto3==1.18.58
botocore==1.21.58
databases==0.5.2
greenlet==1.1.2
idna==3.2
importlib-metadata==4.8.1
jmespath==0.10.0
orm==0.2.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.25
typesystem==0.3.0
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
